// src/app/modules/m3d/effects/anaglyph-effect.d.ts
import { Camera, Matrix3, Scene, WebGLRenderer } from 'three';

export class AnaglyphEffect {
  constructor(renderer: WebGLRenderer, width?: number, height?: number);
  colorMatrixLeft: Matrix3;
  colorMatrixRight: Matrix3;

  dispose(): void;
  render(scene: Scene, camera: Camera): void;
  setSize(width: number, height: number): void;
}